requirements.txt
flask==2.3.2
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.0.0
secure-smtplib==0.1.1
werkzeug==2.3.6
gunicorn==21.2.0